+++ src/categories/routers.py
@@ -3,6 +3,8 @@ from typing import Optional
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from src.auth.auth_bearer import JWTBearer
+from src.auth.services import check_permission
 from src.database import get_async_session
 from src.models import Category
 from src.categories import schemas
@@ -18,26 +20,36 @@ categories_service = CategoryService(Category)
 
 @router.get("/")
 async def get_categories(page: int = None, limit: int = None, search: str = None,
-                         session: AsyncSession = Depends(get_async_session)):
+                         session: AsyncSession = Depends(get_async_session),
+                         current_user: str = Depends(JWTBearer())):
+    check_permission("read_category", current_user)
     return await categories_service.get_entities(session, page, limit, search)
 
 
 @router.post("/")
-async def create_category(category: schemas.CategoryCreate, session: AsyncSession = Depends(get_async_session)):
+async def create_category(category: schemas.CategoryCreate, session: AsyncSession = Depends(get_async_session),
+                          current_user: str = Depends(JWTBearer())):
+    check_permission("create_category", current_user)
     return await categories_service.create_entity(category, session)
 
 
 @router.get("/{category_id}")
-async def get_category(category_id: int, session: AsyncSession = Depends(get_async_session)):
+async def get_category(category_id: int, session: AsyncSession = Depends(get_async_session),
+                       current_user: str = Depends(JWTBearer())):
+    check_permission("read_category", current_user)
     return await categories_service.get_entity(category_id, session)
 
 
 @router.patch("/{category_id}")
 async def update_category(category_id: int, category: schemas.CategoryUpdate,
-                          session: AsyncSession = Depends(get_async_session)):
+                          session: AsyncSession = Depends(get_async_session),
+                          current_user: str = Depends(JWTBearer())):
+    check_permission("update_category", current_user)
     return await categories_service.update_entity(category_id, category, session)
 
 
 @router.delete("/")
-async def delete_category(category_ids: list[int], session: AsyncSession = Depends(get_async_session)):
+async def delete_category(category_ids: list[int], session: AsyncSession = Depends(get_async_session),
+                          current_user: str = Depends(JWTBearer())):
+    check_permission("delete_category", current_user)
     return await categories_service.delete_entities(category_ids, session)
